AR Split Works Now
Raymond and I messed around with the code. Turns out the refreshGallery and resetcam was causing the issue of null pointer in the take picture function. After moving it to the outside of the if statement, the AR function is working as intended and Static function is working as intended. Works on both simulator Pixel API 30 and College given Android Device.